Combine Near2Far with monitor data, put near2far.py into components
