+++ tidy3d/components/data/sim_data.py
@@ -1,13 +1,12 @@
 """ Simulation Level Data """
 from __future__ import annotations
-from typing import Dict, Callable, Union
+from typing import Dict, Callable
 
 import xarray as xr
 import pydantic as pd
 import numpy as np
 
-from .monitor_data import MonitorDataTypes, AbstractFieldData
-from .monitor_data_n2f import Near2FarDataTypes
+from .monitor_data import MonitorDataTypes, MonitorDataType, AbstractFieldData
 from ..base import Tidy3dBaseModel
 from ..simulation import Simulation
 from ..boundary import BlochBoundary
@@ -15,7 +14,6 @@ from ..types import Ax, Axis, annotate_type, Literal
 from ..viz import equal_aspect, add_ax_if_none
 from ...log import DataError, log
 
-MonitorDataType = Union[MonitorDataTypes + Near2FarDataTypes]
 DATA_TYPE_MAP = {data.__fields__["monitor"].type_: data for data in MonitorDataTypes}
